Move struct `Sha3CodeGen` and function `rand_bytes` to mock crate.
