+++ bus-mapping/Cargo.toml
@@ -16,7 +16,6 @@ halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.gi
 itertools = "0.10"
 lazy_static = "1.4"
 log = "0.4.14"
-rand = { version = "0.8", optional = true }
 serde = {version = "1.0.130", features = ["derive"] }
 serde_json = "1.0.66"
 strum = "0.24"
@@ -30,7 +29,6 @@ pretty_assertions = "1.0.0"
 tokio = { version = "1.13", features = ["macros"] }
 url = "2.2.2"
 mock = { path = "../mock" }
-rand = "0.8"
 
 [features]
-test = ["mock", "rand"]
+test = ["mock"]
