+++ bus-mapping/src/evm.rs
@@ -4,5 +4,3 @@ pub(crate) mod opcodes;
 
 pub use eth_types::evm_types::opcode_ids::OpcodeId;
 pub use opcodes::Opcode;
-
-pub use opcodes::Sha3CodeGen;
